+++ src/main/java/se/bjurr/prnfs/admin/AdminFormValues.java
@@ -20,6 +20,10 @@ public final class AdminFormValues extends ArrayList<Map<String, String>> {
   BUTTON_CONFIG_FORM, TRIGGER_CONFIG_FORM, GLOBAL_SETTINGS
  };
 
+ public enum TRIGGER_IF_MERGE {
+  ALWAYS, NOT_CONFLICTING, CONFLICTING
+ };
+
  public enum BUTTON_VISIBILITY {
   NONE, SYSTEM_ADMIN, ADMIN, EVERYONE
  };
@@ -47,6 +51,7 @@ public final class AdminFormValues extends ArrayList<Map<String, String>> {
   admin_allowed, //
   user_allowed, //
   injection_url, //
-  injection_url_regexp
+  injection_url_regexp, //
+  trigger_if_isconflicting
  }
 }