+++ src/main/java/se/bjurr/prnfs/settings/PrnfsNotification.java
@@ -12,6 +12,7 @@ import static se.bjurr.prnfs.admin.AdminFormValues.VALUE;
 import static se.bjurr.prnfs.admin.AdminFormValues.FIELDS.filter_regexp;
 import static se.bjurr.prnfs.admin.AdminFormValues.FIELDS.filter_string;
 import static se.bjurr.prnfs.admin.AdminFormValues.FORM_TYPE.TRIGGER_CONFIG_FORM;
+import static se.bjurr.prnfs.admin.AdminFormValues.TRIGGER_IF_MERGE.ALWAYS;
 import static se.bjurr.prnfs.listener.UrlInvoker.HTTP_METHOD.GET;
 import static se.bjurr.prnfs.settings.PrnfsPredicates.predicate;
 
@@ -22,6 +23,7 @@ import java.util.Map;
 import se.bjurr.prnfs.admin.AdminFormValues;
 import se.bjurr.prnfs.admin.AdminFormValues.FIELDS;
 import se.bjurr.prnfs.admin.AdminFormValues.FORM_TYPE;
+import se.bjurr.prnfs.admin.AdminFormValues.TRIGGER_IF_MERGE;
 import se.bjurr.prnfs.listener.PrnfsPullRequestAction;
 import se.bjurr.prnfs.listener.UrlInvoker.HTTP_METHOD;
 
@@ -44,11 +46,12 @@ public class PrnfsNotification {
  private final String name;
  private final String injectionUrl;
  private final String injectionUrlRegexp;
+ private final TRIGGER_IF_MERGE triggerIfCanMerge;
 
  public PrnfsNotification(List<PrnfsPullRequestAction> triggers, String url, String user, String password,
    String filterString, String filterRegexp, String method, String postContent, List<Header> headers, String proxyUser,
    String proxyPassword, String proxyServer, String proxyPort, String name, String injectionUrl,
-   String injectionUrlRegexp) throws ValidationException {
+   String injectionUrlRegexp, String triggerIfCanMerge) throws ValidationException {
   this.proxyUser = emptyToNull(nullToEmpty(proxyUser).trim());
   this.proxyPassword = emptyToNull(nullToEmpty(proxyPassword).trim());
   this.proxyServer = emptyToNull(nullToEmpty(proxyServer).trim());
@@ -56,6 +59,8 @@ public class PrnfsNotification {
   this.headers = checkNotNull(headers);
   this.postContent = emptyToNull(nullToEmpty(postContent).trim());
   this.method = HTTP_METHOD.valueOf(firstNonNull(emptyToNull(nullToEmpty(method).trim()), GET.name()));
+  this.triggerIfCanMerge = TRIGGER_IF_MERGE.valueOf(firstNonNull(emptyToNull(nullToEmpty(triggerIfCanMerge).trim()),
+    ALWAYS.name()));
   if (nullToEmpty(url).trim().isEmpty()) {
    throw new ValidationException(FIELDS.url.name(), "URL not set!");
   }
@@ -84,6 +89,10 @@ public class PrnfsNotification {
   this.name = firstNonNull(emptyToNull(nullToEmpty(name).trim()), DEFAULT_NAME);
   this.injectionUrl = emptyToNull(nullToEmpty(injectionUrl).trim());
   this.injectionUrlRegexp = emptyToNull(nullToEmpty(injectionUrlRegexp).trim());
+ }
+
+ public TRIGGER_IF_MERGE getTriggerIfCanMerge() {
+  return triggerIfCanMerge;
  }
 
  public Optional<String> getFilterRegexp() {